Rename module paths in Containers
