+++ src/lib/Containers/__stories__/MyAccount.story.tsx
@@ -2,8 +2,8 @@ import { storiesOf } from "@storybook/react-native"
 import * as React from "react"
 import { RootContainer } from "react-relay"
 
-import Routes from "../../relay/routes"
-import MyAccount from "../my_account"
+import Routes from "../../Relay/Routes"
+import MyAccount from "../MyAccount"
 
 storiesOf("My Account").add("Root", () => {
   const profileRoute = new Routes.MyAccount()
